Skeleton blog post for GSoC large project wrap-up (#417)
* Skeleton blog post for GSoC large project wrap-up

@lunamorrow - We would love to share the great work you did during the 2024 GSoC program. Would you both be comfortable sharing an .md file of your final project submission to be posted on the MDAnalysis blog and linked in this post?

If so, it'd be great if you could please add your final GSoC submission report as a blog post as indicated in the README for this repo: https://github.com/MDAnalysis/MDAnalysis.github.io#blog-mdanalysisorgblog and add the appropriate file names to the bottom of the page in the skeleton provided here. As you did for the introductory blog post PR, you'll want to fork the GSoC-lg branch in my github repo.

Let me know if you have any questions. Thanks!

* Fix link formatting

* Add blog post and update reference (#33)

* Add blog post and update reference

* Update _posts/2025-02-09-Luna-Morrow-Final-Submission.md

* Update _posts/2025-02-09-Luna-Morrow-Final-Submission.md

* Update _posts/2025-02-09-Luna-Morrow-Final-Submission.md

* Update _posts/2025-02-09-Luna-Morrow-Final-Submission.md

---------

Co-authored-by: Jenna M Swarthout Goddard <38287809+jennaswa@users.noreply.github.com>

* Update 2025-01-04-gsoc2024-final-lg.md

* Rename 2025-01-04-gsoc2024-final-lg.md to 2025-02-11-gsoc2024-final-lg.md

---------

Co-authored-by: Luna Morrow <159094759+lunamorrow@users.noreply.github.com>